Created utils.h for Assert and ArrayLength and cleaned up platform layer Other misc changes:  - Fixed include guard conventions  - Added asserts for invalid instruction options (I'm pretty sure)  - Added missing casts

+++ code/x86_codegen.cpp
@@ -56,7 +56,7 @@ void GenInstructionsArcList(nfa_arc_list *ArcList, mem_arena *Arena) {
         nfa_transition *Arc = &ArcList->Transitions[TransitionIdx];
 
         if (Arc->From != DisableState) {
-            if (DisableState != -1) {
+            if (DisableState != (uint32_t)-1) {
                 GenInstructionsTransitionSet(DisableState, ActivateMask, Arena);
             }
 
@@ -67,7 +67,7 @@ void GenInstructionsArcList(nfa_arc_list *ArcList, mem_arena *Arena) {
         }
     }
 
-    if (DisableState != -1) {
+    if (DisableState != (uint32_t)-1) {
         GenInstructionsTransitionSet(DisableState, ActivateMask, Arena);
     }
 }
